[Refactor] refactor plain Java code to use JPA
