+++ src/main/java/com/capstone/favicon/application/repository/AuthService.java
@@ -1,0 +1,8 @@
+package com.capstone.favicon.application.repository;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface AuthService {
+    void login(String username, String password, HttpServletRequest request);
+    void logout(HttpServletRequest request);
+}
