+++ src/main/java/com/capstone/favicon/domain/domain/User.java
@@ -1,0 +1,22 @@
+package com.capstone.favicon.domain;
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+
+
+@Getter
+@Setter
+@Entity
+@Table(name = "users")
+public class User {
+
+    @Id
+    private String username;
+    private String password;
+
+}
