+++ src/main/java/com/capstone/favicon/domain/dto/LoginDto.java
@@ -1,0 +1,11 @@
+package com.capstone.favicon.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginDto {
+    private String username;
+    private String password;
+}
