+++ src/main/java/com/capstone/favicon/infrastructure/UserRepository.java
@@ -1,0 +1,12 @@
+package com.capstone.favicon.repository;
+
+import com.capstone.favicon.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+    User findByUsername(String username);
+}
+
