--- src/main/java/com/favicon/capstone/CapstoneApplication.java
@@ -1,13 +1,0 @@
-package com.favicon.capstone;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CapstoneApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CapstoneApplication.class, args);
-	}
-
-}
