--- src/main/java/com/favicon/capstone/user/dto/UserDto.java
@@ -1,11 +1,0 @@
-package com.favicon.capstone.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserDto {
-    private String username;
-    private String password;
-}
