--- src/main/java/com/favicon/capstone/user/service/AuthService.java
@@ -1,33 +1,0 @@
-package com.favicon.capstone.user.service;
-
-import com.favicon.capstone.user.domain.User;
-import com.favicon.capstone.user.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.stereotype.Service;
-import java.util.Collections;
-
-@Service
-@RequiredArgsConstructor
-public class AuthService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    // UserDetailsService의 메서드 구현
-    @Override
-    public UserDetails loadUserByUsername(String id) throws UsernameNotFoundException {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
-        return new org.springframework.security.core.userdetails.User(
-                user.getId(),
-                user.getPassword(),
-                Collections.singleton(new SimpleGrantedAuthority("ROLE_USER"))
-        );
-    }
-}
-
-
-
