--- src/main/resources/application.yml
@@ -1,16 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:postgresql://localhost:5432/favicon
-    username: postgres
-    password: '0204'
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-  thymeleaf:
-    prefix: classpath:/templates/
-    suffix: .html
